Refactor #1
- Move browse view into its own file - this also involved moving MAX_RESULTS_PER_PAGE, as_integer, and paginator into shared utilities.
