+++ judgments/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path, re_path, register_converter
 
 from . import converters, feeds, views
+from .views.browse import browse
 
 register_converter(converters.YearConverter, "yyyy")
 register_converter(converters.DateConverter, "date")
@@ -8,13 +9,13 @@ register_converter(converters.CourtConverter, "court")
 register_converter(converters.SubdivisionConverter, "subdivision")
 
 urlpatterns = [
-    path("<court:court>", views.browse, name="browse"),
-    path("<yyyy:year>", views.browse, name="browse"),
-    path("<court:court>/<yyyy:year>", views.browse, name="browse"),
-    path("<court:court>/<subdivision:subdivision>", views.browse, name="browse"),
+    path("<court:court>", browse, name="browse"),
+    path("<yyyy:year>", browse, name="browse"),
+    path("<court:court>/<yyyy:year>", browse, name="browse"),
+    path("<court:court>/<subdivision:subdivision>", browse, name="browse"),
     path(
         "<court:court>/<subdivision:subdivision>/<yyyy:year>",
-        views.browse,
+        browse,
         name="browse",
     ),
     path("atom.xml", feeds.LatestJudgmentsFeed(), name="feed"),
